Ensure no files are returned if no paths are checked
Prior to this change, if we passed an empty sequence to `ls_files`, we
would run `git ls-files --`, which would return *all* of the files in
the repository, which is the complete opposite of the expected
behaviour. This happened when a non-existent path was present in the
`CODEOWNERS` file.

This change adds a test for the circumstance and fixes the bug.

Co-authored-by: Nick Pope <nick@nickpope.me.uk>