Add diff current (non-checked in) codeowners file
Prior to this change, there was no way to diff non-checked in files.

This change adds a simple special value `"CURRENT"` that when used in
place of the head ref value will instead load the codeowners file from
disk regardless of if it has been checked in/committed or not.