Add a max_files_to_print argument to MarkdownPrinter
Prior to this change, the printer would print every file. This could
result in us printing a message which was too large to upload as a
comment to Github.

This change adds a max_files_to_print argument, which truncates the list
of files and prints a message if truncation occurred.